Fix some clippy lints

+++ crates/kernel/src/device.rs
@@ -188,7 +188,7 @@ pub fn init_devices(tree: &DeviceTree<'_>) {
             .unwrap();
         let (gpio_addr, _) = find_device_addr(gpio).unwrap().unwrap();
         let gpio_base = unsafe { map_device(gpio_addr) }.as_ptr();
-        println!("| GPIO controller addr: {:#010x}", gpio_addr as usize);
+        println!("| GPIO controller addr: {:#010x}", gpio_addr);
         println!("| GPIO controller base: {:#010x}", gpio_base as usize);
         let gpio = unsafe { gpio::bcm2711_gpio_driver::init_with_defaults(gpio_base, true) };
         unsafe { GPIO.init(SpinLock::new(gpio)) };
